refactor(agent): extract MCP client helpers into mcp_client.py
